+++ src/datatype/fs_matrix.h
@@ -8,7 +8,7 @@
 	typedef struct _matrix {
 		int width;
 		int height;
-		double* mat;
+		float* mat;
 	} matrix;
 	#define MAT(m,r,c) (m.mat[(r)*m.width+(c)])
 
@@ -19,8 +19,8 @@
 	void mat_print(matrix mat);
 	matrix* mat_plus(matrix* a, matrix* b);
 	matrix* mat_minus(matrix* a, matrix* b);
-	matrix* mat_scala_multiply(matrix* a, double b);
-	matrix* mat_scala_divide(matrix* a, double b);
+	matrix* mat_scala_multiply(matrix* a, float b);
+	matrix* mat_scala_divide(matrix* a, float b);
 	matrix* mat_multiply(matrix* a, matrix* b);
 	matrix* mat_square_multiply(matrix* a, matrix* b);
 	matrix* mat_divide(matrix* a, matrix* b);
